requirements.txt
anthropic
gspread
google-auth
requests
numpy
scipy
tastytrade
